requirements.txt
Django==4.2.7
djangorestframework==3.15.1
drf-spectacular==0.27.2
Pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.4
torch==2.2.2
yolov5